+++ repoze/what/plugins/pylonshq/__init__.py
@@ -21,11 +21,12 @@ Utilities to use :mod:`repoze.what` v1 in a Pylons or TurboGears 2 application.
 # Let's make all the utilities available in this namespace:
 
 from repoze.what.plugins.pylonshq.utils import booleanize_predicates, \
-                                               debooleanize_predicates
+                                               debooleanize_predicates, \
+                                               is_met, not_met
 
 from repoze.what.plugins.pylonshq.protectors import ActionProtector, \
                                                     ControllerProtector
 
 
 __all__ = ['ActionProtector', 'ControllerProtector', 'booleanize_predicates',
-           'debooleanize_predicates']
+           'debooleanize_predicates', 'is_met', 'not_met']