+++ repoze/what/plugins/pylonshq/utils.py
@@ -24,7 +24,37 @@ from pylons import request
 from repoze.what.predicates import Predicate
 
 
-__all__ = ['booleanize_predicates', 'debooleanize_predicates']
+__all__ = ['is_met', 'not_met', 'booleanize_predicates',
+           'debooleanize_predicates']
+
+
+#{ Evaluators
+
+
+def is_met(predicate):
+    """
+    Evaluate the :mod:`repoze.what` ``predicate`` checker and return ``True``
+    if it's met.
+    
+    :param predicate: The :mod:`repoze.what` predicate checker to be evaluated.
+    :return: ``True`` if it's met; ``False`` otherwise.
+    :rtype: bool
+    
+    """
+    return predicate.is_met(request.environ)
+
+
+def not_met(predicate):
+    """
+    Evaluate the :mod:`repoze.what` ``predicate`` checker and return ``False``
+    if it's met.
+    
+    :param predicate: The :mod:`repoze.what` predicate checker to be evaluated.
+    :return: ``False`` if it's met; ``True`` otherwise.
+    :rtype: bool
+    
+    """
+    return not predicate.is_met(request.environ)
 
 
 #{ Booleanizers
@@ -34,6 +64,11 @@ def booleanize_predicates():
     """
     Make :mod:`repoze.what` predicates evaluable without passing the 
     ``environ`` explicitly.
+    
+    .. warning::
+    
+        The use of this function is **strongly discouraged**. Use
+        :func:`is_met` or :func:`not_met` instead.
     
     """
     Predicate.__nonzero__ = lambda self: self.is_met(request.environ)