+++ tests/fixture/pylonsapp.py
@@ -21,7 +21,8 @@ from pylons import response
 from pylons.controllers import WSGIController
 
 from repoze.what.predicates import All, Not, not_anonymous, is_user, in_group
-from repoze.what.plugins.pylonshq import ActionProtector, ControllerProtector
+from repoze.what.plugins.pylonshq import ActionProtector, ControllerProtector,\
+                                         is_met, not_met
 
 from tests.fixture import special_require
 
@@ -102,3 +103,13 @@ class BasicPylonsController(WSGIController):
     def boolean_predicate(self):
         p = not_anonymous()
         return 'The predicate is %s' % bool(p)
+    
+    def is_met_util(self):
+        if is_met(not_anonymous()):
+            return 'You are not anonymous'
+        return 'You are anonymous'
+    
+    def not_met_util(self):
+        if not_met(not_anonymous()):
+            return 'You are anonymous'
+        return 'You are not anonymous'